refactor: split tests into unit and integration suites
- Rename integration tests to *IT and move them per feature package
- Configure failsafe to run *IT tests and surefire for unit tests
- Make jacoco coverage opt-in through the coverage profile
- Extract shared JVM test args into the test.argLine property
- Replace assembly with a shade dist profile for a single reconciler.jar
- Build the prod image with -Pdist and copy reconciler.jar
